+++ gamedata/fbl/legend-generator.json
