+++ gamedata/myz/artifacts.list.json
