+++ gamedata/myz/scrap.list
