+++ generators/ALIENStarGenerator.js
@@ -1,6 +1,6 @@
 const YZGenerator2 = require('./YZGenerator2');
 const AlienWorld = require('./ALIENWorldGenerator');
-const StarData = require('../gamedata/star-generator.json');
+const StarData = require('../gamedata/alien/star-generator.json');
 const Util = require('../utils/Util');
 
 /**
